docs(memory): document BrowseQueryLike fields and tighten the parity tripwire
Five review findings on the Task 2 core mirror + tripwire:

- BrowseQueryLike shipped with no field docs. The named-mirror refactor dropped
  the since/until/now docs that lived on the old inline browse shape, and the
  widened status/kind never carried across the "an EMPTY set matches nothing"
  rule that packages/memory/src/types.ts states. Restore per-field docs on a
  type this branch made public, matching the style of `search` and
  MemoryRecordLike alongside it.

- Identical<BrowsePage, BrowsePageLike> pinned MemoryRecordLike only
  transitively, so a record-only drift failed on the PAGE naming neither type —
  and the obvious way to quiet that is to loosen the page. Assert
  Identical<MemoryRecord, MemoryRecordLike> on its own line.

- The tripwire's worked example cited `readonly namespace?: string`, which the
  next commit then added for real, making the comment false. Use a field that
  stays hypothetical (`tripwireProbe`), verified to fire.

- Drop the fake runtime expect: the it() asserted `true === true` and its title
  said "mutually assignable" for what are now identity checks. Rename to
  memory-contract-parity.contract.ts and use the repo's existing
  test/*.contract.ts idiom (typecheck-only, `void` the bindings) — already
  covered by both packages/testing tsconfigs.

- The bivariance rationale was written out in three places, one of them
  cross-referencing a type eleven lines up the same file. Keep the single
  statement, on `Identical` where it is used.
